Moved out code from commown related to website_sale_commown_shipping
