+++ website_sale_commown_shipping/controllers/__init__.py
@@ -1,0 +1,1 @@
+from . import website_sale
