+++ website_sale_commown_shipping/views/payment_template.xml
@@ -1,0 +1,13 @@
+<odoo>
+
+  <template id="address_on_payment" inherit_id="website_sale.address_on_payment">
+
+    <!-- Add UI elements to emphasize the user needs to check its delivery address -->
+    <xpath expr="//div[@id='shipping_and_billing']/*[1]" position="before">
+      <h3 class="card-title">Adresse de livraison</h3>
+      <h4 class="card-subtitle mb-3 text-danger">Merci de vérifier vos adresses</h4>
+    </xpath>
+
+  </template>
+
+</odoo>
